requirements.txt
requests
openai
qdrant-client
tqdm
python-dotenv
crewai
fitz
langchain